Refactor extract Editor
