+++ ctui/core.py
@@ -1,6 +1,4 @@
-import os
 import socket
-from subprocess import call
 
 from httplib2 import ServerNotFoundError
 
@@ -11,8 +9,9 @@ from ctui.keybindings import Keybindings
 from ctui.memory import MemoryStore
 from ctui.model.contact import Contact
 from ctui.model.google_contact import GoogleContact
-from ctui.textfile import TextFileStore
 from ctui.rdf import RDFStore
+from ctui.textfile import TextFileStore
+from ctui.service.editor import Editor
 
 
 class Core:
@@ -116,10 +115,6 @@ class Core:
         contacts.sort(key=lambda x: x.name)
         return contacts
 
-    """
-    Returns a list of the names of all contacts.
-    """
-
     def get_all_contact_names(self):
         contact_names = self.rdfstore.get_all_contact_names() \
                         + self.textfilestore.get_all_contact_names()
@@ -205,47 +200,3 @@ class Core:
 
     def get_filter_string(self):
         return self.filter_string
-
-
-class Editor:
-    def __init__(self, editor_name):
-        self.editor = os.environ.get('EDITOR', editor_name)
-
-    def add(self, filepath):
-        temp_filepath = filepath + '.tmp'
-
-        try:
-            with open(temp_filepath, 'w') as tf:
-                call([self.editor, tf.name])
-
-            with open(temp_filepath, 'r') as tf:
-                content = tf.read()
-
-            os.remove(temp_filepath)
-
-        except OSError:
-            raise OSError  # TODO
-
-        return content
-
-    def edit(self, filepath):
-        temp_filepath = filepath + '.tmp'
-
-        try:
-            with open(filepath) as f:
-                old_content = f.read()
-
-            with open(temp_filepath, 'w') as tf:
-                tf.write(old_content)
-                tf.flush()
-                call([self.editor, tf.name])
-
-            with open(temp_filepath, 'r') as tf:
-                content = tf.read()
-
-            os.remove(temp_filepath)
-
-        except OSError:
-            raise OSError  # TODO
-
-        return content
